[Python] Move all python code together
Ref #18 & #19.
