+++ python/pyhusky/backend/globalvar.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import zmq
 import sys
+import zmq
 
 class OperationParam:
     data_str = "data"
