+++ python/pyhusky/backend/library/svm.py
@@ -12,31 +12,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
 import cPickle
-from backend.python.globalvar import GlobalVar, GlobalSocket, OperationParam
+import json
 
-import sys
+from pyhusky.backend.globalvar import GlobalVar, GlobalSocket, OperationParam
 
 def register_all():
-    LinearModel.register()
+    SVMModel.register()
 
-class LinearModel:
+class SVMModel:
     @staticmethod
     def register():
-        GlobalVar.name_to_prefunc["LinearRegressionModel#LinearR_load_pyhlist_py"] = LinearModel.load_pyhlist_prefunc
-        GlobalVar.name_to_func["LinearRegressionModel#LinearR_load_pyhlist_py"] = LinearModel.load_pyhlist_func
-        GlobalVar.name_to_func["LinearRegressionModel#LinearR_load_pyhlist_end_py"] = None
-        GlobalVar.name_to_postfunc["LinearRegressionModel#LinearR_load_pyhlist_end_py"] = LinearModel.load_pyhlist_end_postfunc
+        GlobalVar.name_to_prefunc["SVMModel#SVM_load_pyhlist_py"] = SVMModel.load_pyhlist_prefunc
+        GlobalVar.name_to_func["SVMModel#SVM_load_pyhlist_py"] = SVMModel.load_pyhlist_func
+        GlobalVar.name_to_func["SVMModel#SVM_load_pyhlist_end_py"] = None
+        GlobalVar.name_to_postfunc["SVMModel#SVM_load_pyhlist_end_py"] = SVMModel.load_pyhlist_end_postfunc
 
-        GlobalVar.name_to_type["LinearRegressionModel#LinearR_load_pyhlist_py"] = GlobalVar.librarytype
-        GlobalVar.name_to_type["LinearRegressionModel#LinearR_load_pyhlist_end_py"] = GlobalVar.librarytype
+        GlobalVar.name_to_type["SVMModel#SVM_load_pyhlist_py"] = GlobalVar.librarytype
+        GlobalVar.name_to_type["SVMModel#SVM_load_pyhlist_end_py"] = GlobalVar.librarytype
 
     # For those functions that need to load from PyHuskyList
     @staticmethod
     def load_pyhlist_prefunc(op):
         GlobalVar.xy_line_list = op.op_param[OperationParam.list_str]
-        GlobalVar.is_sparse = op.op_param["is_sparse"]
         GlobalVar.xy_line_store = []
     @staticmethod
     def load_pyhlist_func(op, data):
@@ -47,9 +45,8 @@ class LinearModel:
             GlobalVar.xy_line_store.append(I)
     @staticmethod
     def load_pyhlist_end_postfunc(op):
-        GlobalSocket.pipe_to_cpp.send("LinearRegressionModel#LinearR_load_pyhlist_py")
+        GlobalSocket.pipe_to_cpp.send("SVMModel#SVM_load_pyhlist_py")
         GlobalSocket.pipe_to_cpp.send(GlobalVar.xy_line_list)
-        GlobalSocket.pipe_to_cpp.send(GlobalVar.is_sparse)
         # send out the len of datatset
         GlobalSocket.pipe_to_cpp.send(str(len(GlobalVar.xy_line_store)))
         for (X, y) in GlobalVar.xy_line_store:
