+++ python/pyhusky/common/operation.py
@@ -29,6 +29,6 @@ class Operation:
         self.op_param = op_param
         self.op_deps = [dep for dep in op_deps if dep is not None]
         self.is_materialized = False
+
     def __repr__(self):
         return "<"+self.op_name+">"
-        # return "<"+self.op_name+">"+" ".join(str(x) for x in self.op_param)
